+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## dev
 
+* Improved error messages concerning failed HTTP requests.
 * Make the OS name used in the default user-agent string more readable (e.g.
   `freebsd` -> `FreeBSD`).